08_django/11_DRF/10-02-django-rest-framework-skeleton/requirements.txt
asgiref==3.7.2
attrs==23.1.0
Django==4.2.2
djangorestframework==3.14.0
drf-spectacular==0.26.5
inflection==0.5.1
jsonschema==3.2.0
pyrsistent==0.19.3
pytz==2023.3
PyYAML==6.0.1
six==1.16.0
spectacular==0.4.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
